+++ whatToWatch/src/com/team2/wtw/faq/ProcessQA.java
@@ -1,0 +1,13 @@
+package com.team2.wtw.faq;
+
+import main.PlayAllFunction;
+
+public class ProcessQA {
+
+public static void main(String[] args) throws Exception {
+		
+		PlayAllFunction paf = new PlayAllFunction();
+		paf.goFunction();
+	}
+	
+}
