--- whatToWatch/src/main/Main.java
@@ -1,17 +1,0 @@
-package main;
-
-import java.util.Scanner;
-
-import playstart.PlayStart;
-
-public class Main {
-	public static final Scanner SC = new Scanner(System.in);
-	
-	public static void main(String[] args) throws Exception {
-		
-		PlayAllFunction paf = new PlayAllFunction();
-		paf.goFunction();
-	}
-
-}
-//회원번호받아오기..
